debug

+++ uitls_grad_eval.py
@@ -51,7 +51,7 @@ def _pad_batch(batch: Sequence[Sequence[Any]]) -> TYPE_BATCH:
         if total_video_len > max_video_len:
             max_video_len = total_video_len
 
-    text_tensor = torch.zeros(batch_size, max_text_len, dtype=torch.long, requires_grad=True)
+    text_tensor = torch.zeros(batch_size, max_text_len, dtype=torch.float, requires_grad=True)
     video_tensor = torch.zeros(batch_size, max_video_len, video.shape[1], dtype=torch.float, requires_grad=True)
